Refactor components name
