+++ src/components/Card/index.js
@@ -1,3 +1,3 @@
 export { default as Cards } from './Cards';
-export { default as Edit } from './Edit';
+export { default as EditButton } from './EditButton';
 export { default as Title } from './Title';
