+++ src/components/List/index.js
@@ -1,1 +1,2 @@
-export { default } from './Lists';
+export { default as Lists } from './Lists';
+export { default as Title } from './Title';
