+++ src/components/Shared/index.js
@@ -1,5 +1,4 @@
 export { default as InputSet } from './InputSet';
-export { default as Title } from './Title';
 export { default as MenuButton } from './MenuButton';
 export { default as EditButton } from './EditButton';
 export { default as DeleteButton } from './DeleteButton';
